src/Flytest/requirements.txt
numpy==1.15.4
opencv-python==3.4.3.18
pygame==1.9.4
tensorflow
matplotlib
Pillow
argparse
